doesn't work the new guard client, so I'm gonna use the default use

+++ database/migrations/2020_12_14_191433_create_clients_table.php
@@ -19,7 +19,7 @@ class CreateClientsTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('telephone');
+            $table->bigInteger('telephone');
             $table->rememberToken();
             $table->timestamps();
         });
